+++ app/Elca/Model/ProcessConfig/Conversion/RequiredConversion.php
@@ -25,6 +25,7 @@
 
 namespace Elca\Model\ProcessConfig\Conversion;
 
+use Elca\Model\Common\Quantity\Quantity;
 use Elca\Model\Common\Unit;
 use Elca\Model\Exception\InvalidArgumentException;
 use Utils\Model\SurrogateIdTrait;